+++ MapboxSearch/base/src/test/java/com/mapbox/search/base/AsyncOperationTaskTest.kt
@@ -3,7 +3,6 @@ package com.mapbox.search.base
 import com.mapbox.search.base.task.AsyncOperationTaskImpl
 import com.mapbox.search.base.task.ExtendedAsyncOperationTask
 import com.mapbox.test.dsl.TestCase
-import io.mockk.every
 import io.mockk.mockk
 import org.junit.jupiter.api.TestFactory
 import java.util.concurrent.Future
@@ -17,12 +16,22 @@ internal class AsyncOperationTaskTest {
             When("CompletedAsyncOperationTask is in initial state") {
                 Then("Task is always done", true, task.isDone)
                 Then("Task is not cancelled", false, task.isCancelled)
+                Then("Callback delegate is null", null, task.callbackDelegate)
             }
 
             When("Call CompletedAsyncOperationTask cancel()") {
+                val callback = mockk<() -> Unit>(relaxed = true)
+                task.onCancelCallback = callback
+
                 task.cancel()
+
                 Then("Task is always done", true, task.isDone)
                 Then("Task is not cancelled", false, task.isCancelled)
+                Then("Callback delegate is null", null, task.callbackDelegate)
+
+                VerifyNo("cancel() not on a wrapped object") {
+                    callback()
+                }
             }
         }
     }
@@ -30,28 +39,158 @@ internal class AsyncOperationTaskTest {
     @TestFactory
     fun `Check AsyncOperationTaskImpl`() = TestCase {
         Given("AsyncOperationTaskImpl instance") {
-
-            fun mockOnCancelledCallback(): () -> Unit {
-                val callback = mockk<() -> Unit>(relaxed = true)
-                every { callback() } returns Unit
-                return callback
-            }
+            val delegate = Runnable { }
 
             When("Object is in initial state") {
                 val task = AsyncOperationTaskImpl<Any>()
 
                 Then("Task is not cancelled", false, task.isCancelled)
                 Then("Task is not done", false, task.isDone)
+                Then("callbackActionExecuted is false", false, task.callbackActionExecuted)
+                Then("Callback delegate is null", null, task.callbackDelegate)
+                Then("onCancelCallback is null", null, task.onCancelCallback)
+            }
+
+            When("Object with delegate and onCancelCallback is in initial state") {
+                val task = AsyncOperationTaskImpl<Any>(delegate)
+
+                val onCancelCallback = mockk<() -> Unit>(relaxed = true)
+                task.onCancelCallback = onCancelCallback
+
+                Then("Task is not cancelled", false, task.isCancelled)
+                Then("Task is not done", false, task.isDone)
+                Then("callbackActionExecuted is false", false, task.callbackActionExecuted)
+                Then("Callback delegate is set", delegate, task.callbackDelegate)
+                Then("onCancelCallback delegate is set", onCancelCallback, task.onCancelCallback)
+            }
+
+            When("onComplete() called") {
+                val task = AsyncOperationTaskImpl<Any>(delegate)
+
+                val onCancelCallback = mockk<() -> Unit>(relaxed = true)
+                task.onCancelCallback = onCancelCallback
+
+                task.onComplete()
+
+                Then("Task is not cancelled", false, task.isCancelled)
+                Then("Task is done", true, task.isDone)
+                Then("Callback delegate is null", null, task.callbackDelegate)
+                Then("onCancelCallback is null", null, task.onCancelCallback)
+
+                VerifyNo("OnCancelled callback is not called") {
+                    onCancelCallback()
+                }
+            }
+
+            When("markExecutedAndRunOnCallback() called") {
+                val task = AsyncOperationTaskImpl<Any>(delegate)
+
+                val onCancelCallback = mockk<() -> Unit>(relaxed = true)
+                task.onCancelCallback = onCancelCallback
+
+                val action: Any.() -> Unit = mockk(relaxed = true)
+                task.markExecutedAndRunOnCallback(action)
+
+                Then("Task is not cancelled", false, task.isCancelled)
+                Then("Task is done", true, task.isDone)
+                Then("callbackActionExecuted is true", true, task.callbackActionExecuted)
+                Then("Callback delegate is null", null, task.callbackDelegate)
+                Then("onCancelCallback is null", null, task.onCancelCallback)
+
+                VerifyOnce("Action called") {
+                    action(delegate)
+                }
+
+                VerifyNo("OnCancelled callback is not called") {
+                    onCancelCallback()
+                }
+            }
+
+            When("markExecutedAndRunOnCallback() called on a completed object") {
+                val task = AsyncOperationTaskImpl<Any>(delegate)
+                task.onComplete()
+
+                val action: Any.() -> Unit = mockk(relaxed = true)
+                task.markExecutedAndRunOnCallback(action)
+
+                Then("Task is not cancelled", false, task.isCancelled)
+                Then("Task is done", true, task.isDone)
+                Then("callbackActionExecuted is false", false, task.callbackActionExecuted)
+                Then("Callback delegate is null", null, task.callbackDelegate)
+                Then("onCancelCallback is null", null, task.onCancelCallback)
+
+                VerifyNo("Action is not called") {
+                    action(any())
+                }
+            }
+
+            When("markExecutedAndRunOnCallback() called on a cancelled object") {
+                val task = AsyncOperationTaskImpl<Any>(delegate)
+                task.cancel()
+
+                val action: Any.() -> Unit = mockk(relaxed = true)
+                task.markExecutedAndRunOnCallback(action)
+
+                Then("Task is not cancelled", true, task.isCancelled)
+                Then("Task is done", false, task.isDone)
+                Then("callbackActionExecuted is false", false, task.callbackActionExecuted)
+                Then("Callback delegate is null", null, task.callbackDelegate)
+                Then("onCancelCallback is null", null, task.onCancelCallback)
+
+                VerifyNo("Action is not called") {
+                    action(any())
+                }
+            }
+
+            When("onComplete() called on a cancelled object") {
+                val task = AsyncOperationTaskImpl<Any>(delegate)
+
+                val onCancelCallback = mockk<() -> Unit>(relaxed = true)
+                task.onCancelCallback = onCancelCallback
+
+                task.cancel()
+
+                task.onComplete()
+
+                Then("Task is cancelled", true, task.isCancelled)
+                Then("Task is not done", false, task.isDone)
+                Then("callbackActionExecuted is false", false, task.callbackActionExecuted)
+                Then("Callback delegate is null", null, task.callbackDelegate)
+                Then("onCancelCallback is null", null, task.onCancelCallback)
+
+                VerifyOnce("Callback called once") {
+                    onCancelCallback()
+                }
+            }
+
+            When("onComplete() called more than once") {
+                val task = AsyncOperationTaskImpl<Any>(delegate)
+
+                val onCancelCallback = mockk<() -> Unit>(relaxed = true)
+                task.onCancelCallback = onCancelCallback
+
+                task.onComplete()
+                task.onComplete()
+
+                Then("Task is not cancelled", false, task.isCancelled)
+                Then("Task is done", true, task.isDone)
+                Then("callbackActionExecuted is false", false, task.callbackActionExecuted)
+                Then("Callback delegate is null", null, task.callbackDelegate)
+                Then("onCancelCallback is null", null, task.onCancelCallback)
+
+                VerifyNo("onCancelCallback is not called") {
+                    onCancelCallback()
+                }
             }
 
             When("Object cancelled") {
-                val task = AsyncOperationTaskImpl<Any>()
+                val task = AsyncOperationTaskImpl<Any>(delegate)
 
                 val wrappedTask = mockk<Future<*>>(relaxed = true)
                 task += wrappedTask
 
-                val callback = mockOnCancelledCallback()
-                task.onCancelCallback = callback
+                val onCancelCallback = mockk<() -> Unit>(relaxed = true)
+                task.onCancelCallback = onCancelCallback
 
                 task.cancel()
 
@@ -60,57 +199,86 @@ internal class AsyncOperationTaskTest {
                 }
 
                 VerifyOnce("onCancelCallback called") {
-                    callback()
+                    onCancelCallback()
                 }
 
                 Then("Task is cancelled", true, task.isCancelled)
                 Then("Task is not done", false, task.isDone)
-                Then("onCancelCallback cleared", null, task.onCancelCallback)
+                Then("callbackActionExecuted is false", false, task.callbackActionExecuted)
+                Then("Callback delegate is null", null, task.callbackDelegate)
+                Then("OnCancel callback is null", null, task.onCancelCallback)
 
-                task.onCancelCallback = callback
-
+                task.onCancelCallback = onCancelCallback
                 Then("onCancelCallback can't be set on a cancelled object", null, task.onCancelCallback)
+
+                task.callbackDelegate = delegate
+                Then("Callback delegate can't be set on a cancelled object", null, task.callbackDelegate)
             }
 
-            When("onComplete() called on a cancelled object") {
-                val task = AsyncOperationTaskImpl<Any>()
+            When("markCancelledAndRunOnCallback() called") {
+                val task = AsyncOperationTaskImpl<Any>(delegate)
 
-                val callback = mockOnCancelledCallback()
-                task.onCancelCallback = callback
+                val onCancelCallback = mockk<() -> Unit>(relaxed = true)
+                task.onCancelCallback = onCancelCallback
 
-                task.cancel()
+                val action: Any.() -> Unit = mockk(relaxed = true)
+                task.markCancelledAndRunOnCallback(action)
 
-                task.onComplete()
+                Then("Task is not cancelled", true, task.isCancelled)
+                Then("Task is done", false, task.isDone)
+                Then("callbackActionExecuted is true", true, task.callbackActionExecuted)
+                Then("Callback delegate is null", null, task.callbackDelegate)
+                Then("onCancelCallback is null", null, task.onCancelCallback)
 
-                Then("Task is cancelled", true, task.isCancelled)
-                Then("Task is not done", false, task.isDone)
+                VerifyOnce("Action called") {
+                    action(delegate)
+                }
 
-                VerifyOnce("Callback is not called again") {
-                    callback()
+                VerifyOnce("onCancelCallback callback once") {
+                    onCancelCallback()
                 }
             }
 
-            When("onComplete() called more than once") {
-                val task = AsyncOperationTaskImpl<Any>()
-
-                val callback = mockOnCancelledCallback()
-                task.onCancelCallback = callback
-
+            When("markCancelledAndRunOnCallback() called on a completed object") {
+                val task = AsyncOperationTaskImpl<Any>(delegate)
                 task.onComplete()
-                task.onComplete()
+
+                val action: Any.() -> Unit = mockk(relaxed = true)
+                task.markCancelledAndRunOnCallback(action)
 
                 Then("Task is not cancelled", false, task.isCancelled)
                 Then("Task is done", true, task.isDone)
+                Then("callbackActionExecuted is false", false, task.callbackActionExecuted)
+                Then("Callback delegate is null", null, task.callbackDelegate)
+                Then("onCancelCallback is null", null, task.onCancelCallback)
 
-                VerifyNo("Callback is not called") {
-                    callback()
+                VerifyNo("Action is not called") {
+                    action(any())
+                }
+            }
+
+            When("markCancelledAndRunOnCallback() called on a cancelled object") {
+                val task = AsyncOperationTaskImpl<Any>(delegate)
+                task.cancel()
+
+                val action: Any.() -> Unit = mockk(relaxed = true)
+                task.markCancelledAndRunOnCallback(action)
+
+                Then("Task is not cancelled", true, task.isCancelled)
+                Then("Task is done", false, task.isDone)
+                Then("callbackActionExecuted is false", false, task.callbackActionExecuted)
+                Then("Callback delegate is null", null, task.callbackDelegate)
+                Then("onCancelCallback is null", null, task.onCancelCallback)
+
+                VerifyNo("Action is not called") {
+                    action(any())
                 }
             }
 
             When("cancel() called on a completed task") {
                 val task = AsyncOperationTaskImpl<Any>()
 
-                val callback = mockOnCancelledCallback()
+                val callback = mockk<() -> Unit>(relaxed = true)
                 task.onCancelCallback = callback
 
                 val wrappedTask = mockk<Future<*>>(relaxed = true)
@@ -121,6 +289,9 @@ internal class AsyncOperationTaskTest {
 
                 Then("Task is not cancelled", false, task.isCancelled)
                 Then("Task is done", true, task.isDone)
+                Then("callbackActionExecuted is false", false, task.callbackActionExecuted)
+                Then("Callback delegate is null", null, task.callbackDelegate)
+                Then("OnCancel callback is null", null, task.onCancelCallback)
 
                 VerifyNo("Callback is not called") {
                     callback()
@@ -134,7 +305,7 @@ internal class AsyncOperationTaskTest {
             When("Inner task added to a canceled task") {
                 val task = AsyncOperationTaskImpl<Any>()
 
-                val callback = mockOnCancelledCallback()
+                val callback = mockk<() -> Unit>(relaxed = true)
                 task.onCancelCallback = callback
 
                 task.cancel()
@@ -144,6 +315,8 @@ internal class AsyncOperationTaskTest {
 
                 Then("Task is cancelled", true, task.isCancelled)
                 Then("Task is not done", false, task.isDone)
+                Then("Callback delegate is null", null, task.callbackDelegate)
+                Then("OnCancel callback is null", null, task.onCancelCallback)
 
                 VerifyOnce("Callback called once") {
                     callback()
@@ -157,7 +330,7 @@ internal class AsyncOperationTaskTest {
             When("Inner task added to a completed task") {
                 val task = AsyncOperationTaskImpl<Any>()
 
-                val callback = mockOnCancelledCallback()
+                val callback = mockk<() -> Unit>(relaxed = true)
                 task.onCancelCallback = callback
 
                 task.onComplete()