+++ MapboxSearch/sample/src/main/java/com/mapbox/search/sample/api/CategorySearchJavaExampleActivity.java
@@ -13,7 +13,7 @@ import com.mapbox.search.ResponseInfo;
 import com.mapbox.search.SearchCallback;
 import com.mapbox.search.SearchEngine;
 import com.mapbox.search.SearchEngineSettings;
-import com.mapbox.search.SearchRequestTask;
+import com.mapbox.search.common.AsyncOperationTask;
 import com.mapbox.search.result.SearchResult;
 import com.mapbox.search.sample.BuildConfig;
 
@@ -22,7 +22,7 @@ import java.util.List;
 public class CategorySearchJavaExampleActivity extends AppCompatActivity {
 
     private SearchEngine searchEngine;
-    private SearchRequestTask searchRequestTask;
+    private AsyncOperationTask searchRequestTask;
 
     private final SearchCallback searchCallback = new SearchCallback() {