+++ MapboxSearch/sample/src/main/java/com/mapbox/search/sample/api/JapanSearchJavaExampleActivity.java
@@ -15,8 +15,8 @@ import com.mapbox.search.ResponseInfo;
 import com.mapbox.search.SearchEngine;
 import com.mapbox.search.SearchEngineSettings;
 import com.mapbox.search.SearchOptions;
-import com.mapbox.search.SearchRequestTask;
 import com.mapbox.search.SearchSelectionCallback;
+import com.mapbox.search.common.AsyncOperationTask;
 import com.mapbox.search.result.SearchResult;
 import com.mapbox.search.result.SearchSuggestion;
 import com.mapbox.search.sample.BuildConfig;
@@ -26,7 +26,7 @@ import java.util.List;
 public class JapanSearchJavaExampleActivity extends AppCompatActivity {
 
     private SearchEngine searchEngine;
-    private SearchRequestTask searchRequestTask;
+    private AsyncOperationTask searchRequestTask;
 
     private final SearchSelectionCallback searchCallback = new SearchSelectionCallback() {