+++ MapboxSearch/sample/src/main/java/com/mapbox/search/sample/api/OfflineReverseGeocodingJavaExampleActivity.java
@@ -7,7 +7,6 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 
-import com.mapbox.android.core.location.LocationEngineProvider;
 import com.mapbox.common.Cancelable;
 import com.mapbox.common.TileRegionLoadOptions;
 import com.mapbox.common.TileStore;
@@ -20,11 +19,10 @@ import com.mapbox.search.OfflineSearchEngine.EngineReadyCallback;
 import com.mapbox.search.OfflineSearchEngineSettings;
 import com.mapbox.search.ResponseInfo;
 import com.mapbox.search.SearchCallback;
-import com.mapbox.search.SearchRequestTask;
+import com.mapbox.search.common.AsyncOperationTask;
 import com.mapbox.search.result.SearchResult;
 import com.mapbox.search.sample.BuildConfig;
 
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
@@ -33,7 +31,7 @@ public class OfflineReverseGeocodingJavaExampleActivity extends AppCompatActivit
     private OfflineSearchEngine searchEngine;
     private Cancelable tilesLoadingTask;
     @Nullable
-    private SearchRequestTask searchRequestTask;
+    private AsyncOperationTask searchRequestTask;
 
     private final EngineReadyCallback engineReadyCallback = new EngineReadyCallback() {
         @Override