+++ MapboxSearch/sample/src/main/java/com/mapbox/search/sample/api/ReverseGeocodingKotlinExampleActivity.kt
@@ -10,14 +10,14 @@ import com.mapbox.search.ReverseGeoOptions
 import com.mapbox.search.SearchCallback
 import com.mapbox.search.SearchEngine
 import com.mapbox.search.SearchEngineSettings
-import com.mapbox.search.SearchRequestTask
+import com.mapbox.search.common.AsyncOperationTask
 import com.mapbox.search.result.SearchResult
 import com.mapbox.search.sample.BuildConfig
 
 class ReverseGeocodingKotlinExampleActivity : AppCompatActivity() {
 
     private lateinit var searchEngine: SearchEngine
-    private lateinit var searchRequestTask: SearchRequestTask
+    private lateinit var searchRequestTask: AsyncOperationTask
 
     private val searchCallback = object : SearchCallback {